LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/reporting-subresource-corp.https-expected.txt

CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/nothing-same-origin-corp.txt because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/html/cross-origin-embedder-policy/resources/nothing-same-origin-corp.txt due to access control checks.
CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/common/text-plain.txt because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/common/text-plain.txt due to access control checks.


Harness Error (TIMEOUT), message = null

PASS [document] same-origin
PASS [document] blocked by CORP: same-origin
TIMEOUT [document] blocked due to COEP Test timed out
NOTRUN [document] blocked during redirect
NOTRUN [dedicated worker] same-origin
NOTRUN [dedicated worker] blocked by CORP: same-origin
NOTRUN [dedicated worker] blocked due to COEP
NOTRUN [dedicated worker] blocked during redirect
NOTRUN [shared worker] same-origin
NOTRUN [shared worker] blocked by CORP: same-origin
NOTRUN [shared worker] blocked due to COEP
NOTRUN [shared worker] blocked during redirect
NOTRUN [service worker] same-origin
NOTRUN [service worker] blocked by CORP: same-origin
NOTRUN [service worker] blocked due to COEP
NOTRUN [service worker] blocked during redirect
NOTRUN [between service worker and page] same-origin
NOTRUN [between service worker and page] blocked by CORP: same-origin
NOTRUN [between service worker and page] blocked due to COEP
NOTRUN [between service worker and page] blocked during redirect
NOTRUN destination: script
